add the group chat tab with basic functionality
