+++ app/frontend/src/__mocks__/service.ts
@@ -1,4 +1,6 @@
+import { MessageProps } from "../features/messages/messagelist/Message";
 import { User } from "../pb/api_pb";
+import { Message } from "../pb/conversations_pb";
 import { service as originalService } from "../service";
 import funnycat from "../stories/assets/funnycat.jpg";
 import funnydog from "../stories/assets/funnydog.jpg";
@@ -29,7 +31,7 @@ const userMap = new Map(
   [user1, user2, user3].map((user) => [user.userId, user])
 );
 
-export const mockedService = {
+export const mockedService = ({
   user: {
     getUser: (id: string) => {
       const result = userMap.get(+id);
@@ -37,7 +39,25 @@ export const mockedService = {
     },
   },
   api: { listFriends: () => Promise.resolve([user2.userId, user3.userId]) },
-} as typeof originalService;
+  conversations: {
+    listGroupChats: () =>
+      Promise.resolve([
+        {
+          groupChatId: 3,
+          title: "groupchattitle",
+          memberUserIdsList: [],
+          adminUserIdsList: [],
+          onlyAdminsInvite: true,
+          isDm: false,
+          // created?: google_protobuf_timestamp_pb.Timestamp.AsObject,
+          unseenMessageCount: 0,
+          lastSeenMessageId: 4,
+          latestMessage: message1,
+        },
+      ]),
+    getGroupChatMessages: () => Promise.resolve([message1, message2]),
+  },
+} as unknown) as typeof originalService;
 
 export const service = new Proxy(
   {},
@@ -78,3 +98,16 @@ export const service = new Proxy(
     },
   }
 );
+
+export const message1: MessageProps["message"] = {
+  messageId: 1,
+  authorUserId: 2,
+  text: { text: "testtext" },
+  time: { seconds: Math.floor(+new Date(2020, 0, 1) / 1e3), nanos: 0 },
+} as Message.AsObject;
+const message2: MessageProps["message"] = {
+  messageId: 2,
+  authorUserId: 2,
+  text: { text: "testtext" },
+  time: { seconds: Math.floor(+new Date(2020, 0, 1) / 1e3), nanos: 0 },
+};
